feat(tech): extract CheckList Component
Ref: #5
