+++ src/features/checklist/index.ts
@@ -1,0 +1,3 @@
+import CheckList from "./CheckList.vue";
+
+export { CheckList };
