+++ vite.config.ts
@@ -1,7 +1,13 @@
-import { defineConfig } from 'vite'
-import vue from '@vitejs/plugin-vue'
+import { defineConfig } from "vite";
+import vue from "@vitejs/plugin-vue";
+import path from "path";
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [vue()],
-})
+  resolve: {
+    alias: {
+      "@features": path.resolve(__dirname, "./src/features"),
+    },
+  },
+});
